+++ crates/remote/Cargo.toml
@@ -44,6 +44,7 @@ tracing-error = "0.2"
 thiserror = "2.0.12"
 ts-rs = { git = "https://github.com/xazukx/ts-rs.git", branch = "use-ts-enum", features = ["uuid-impl", "chrono-impl", "no-serde-warnings", "serde-json-impl"] }
 api-types = { path = "../api-types" }
+relay-tunnel = { path = "../../crates/relay-tunnel" }
 utils = { path = "../utils" }
 uuid = { version = "1", features = ["serde", "v4"] }
 jsonwebtoken = { version = "10.2.0", features = ["rust_crypto"] }
