+++ crates/server/Cargo.toml
@@ -17,6 +17,7 @@ git = { path = "../git" }
 git-host = { path = "../git-host" }
 db = { path = "../db" }
 services = { path = "../services" }
+relay-tunnel = { path = "../relay-tunnel" }
 tokio = { workspace = true }
 shlex = "1.3.0"
 tokio-util = { version = "0.7", features = ["io"] }
