+++ crates/utils/src/lib.rs
@@ -19,7 +19,6 @@ pub mod stream_lines;
 pub mod text;
 pub mod tokio;
 pub mod version;
-pub mod ws_io;
 
 /// Cache for WSL2 detection result
 static WSL2_CACHE: OnceLock<bool> = OnceLock::new();
